update: organize services imports by indexing
