+++ src/components/MainContent/About/About.jsx
@@ -1,5 +1,5 @@
 import { faUpRightFromSquare } from "@fortawesome/free-solid-svg-icons";
-import { intersectionService } from "../../../services/intersection-service";
+import { intersectionService } from "../../../services";
 
 import {Navbar} from "../../Navbar";
