+++ src/components/MainContent/Hero/Hero.jsx
@@ -1,4 +1,4 @@
-import { intersectionService } from "../../../services/intersection-service";
+import { intersectionService } from "../../../services";
 import { faArrowDown } from "@fortawesome/free-solid-svg-icons";
 
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
