+++ src/components/MainContent/Projects/Projects.jsx
@@ -1,5 +1,5 @@
 import { faArrowDown } from "@fortawesome/free-solid-svg-icons";
-import { intersectionService } from "../../../services/intersection-service";
+import { intersectionService } from "../../../services";
 
 import { Modal, ProjectModal, ProjectFilter, ProjectCard } from "../../../components";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
